requirements.txt
transformers==4.57.1
datasets==4.4.1
evaluate==0.4.6
wandb==0.22.3
python-dotenv==1.2.1
scipy==1.16.3
scikit-learn==1.7.2